08 Não Permitir Deletar Plano com Detalhes no LaraFood

+++ larafood/app/Http/Controllers/Admin/PlanController.php
@@ -52,11 +52,23 @@ class PlanController extends Controller
 
     public function destroy($url)
     {
-        $plan = $this->repository->where('url', $url)->first();
+        $plan = $this->repository
+                    ->with('details')
+                    ->where('url', $url)
+                    ->first();
+
         if (!$plan) {
             return redirect()->back();
         }
+
+        if($plan->details->count() > 0){
+            return redirect()
+                ->back()
+                ->with('error', 'Já detalhes vinculados a esse plano, Delete os Detalhes antes');
+        }
+
         $plan->delete();
+
         return redirect()->route('plans.index');
     }
